restructure components, remove obsolete/unused styles
