+++ frontend/src/modules/accounts/AccountDetails/AccountDetails.module.scss
@@ -6,46 +6,17 @@
   display: flex;
   flex-direction: column;
   align-self: center;
-  max-width: 1690px;
-  width: 100%;
 
   .header {
     display: flex;
     .detailsCard {
       flex: 3;
     }
-    .actionCard {
-      position: relative;
-      flex: 2;
-      display: flex;
-      align-items: center;
-      padding: $spacing-l $spacing-xl;
-      margin-left: $spacing-base;
-      .avatarWrapper {
-        display: flex;
-        margin-right: $spacing-l * 1.5;
-      }
-      .actionsWrapper, .avatarWrapper {
-        z-index: 1;
-      }
-      .background {
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        z-index: 0;
-        opacity: 0.25;
-      }
-    }
   }
 
-  .currency {
+  .flowCurrency {
     color: $strong-green;
     margin-left: $spacing-s;
   }
 
-  .script {
-    margin-bottom: $spacing-l;
-  }
 }
