+++ frontend/src/modules/contracts/ContractDetails/ContractDetails.module.scss
@@ -3,7 +3,4 @@
 .root {
   display: flex;
   flex-direction: column;
-  align-self: center;
-  max-width: 1690px;
-  width: 100%;
 }
