+++ frontend/src/modules/transactions/SignaturesTable/SignaturesTable.module.scss
@@ -1,0 +1,6 @@
+.hash {
+  max-width: 20vw;
+  min-width: 10vw;
+  display: inline-block;
+  overflow: hidden;
+}
