+++ frontend/src/modules/transactions/TransactionDetails/TransactionDetails.module.scss
@@ -1,4 +1,5 @@
 @import "styles/spacings";
+@import "styles/animations";
 
 .root {
   display: flex;
