+++ frontend/src/modules/transactions/TransactionOverview/TransactionOverview.module.scss
@@ -1,0 +1,5 @@
+.ellipsis {
+  max-width: 13vw;
+  display: inline-block;
+  overflow: hidden;
+}
