--- frontend/src/modules/transactions/details/Details.module.scss
@@ -1,131 +1,0 @@
-@import "styles/spacings";
-@import "styles/animations";
-
-.root {
-  display: flex;
-  flex-direction: column;
-  align-self: center;
-  max-width: 1690px;
-  width: 100%;
-
-  .listCard {
-    display: flex;
-    justify-content: space-between;
-
-    & > div {
-      display: flex;
-      flex-direction: column;
-    }
-  }
-  .hash {
-    max-width: 20vw;
-    min-width: 10vw;
-    display: inline-block;
-    overflow: hidden;
-  }
-
-  .detailsTable {
-    margin-bottom: $spacing-base;
-    position: relative;
-    margin-left: auto;
-    display: flex;
-  }
-
-  .card {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: $spacing-base $spacing-l;
-    margin-bottom: $spacing-base;
-
-    & > div {
-      display: flex;
-      flex-direction: column;
-      padding-right: $spacing-base * 2;
-      flex: 10;
-    }
-
-    & > div:nth-child(2) {
-      flex: 16;
-      overflow-wrap: anywhere;
-    }
-
-    & > div:nth-child(3),
-    & > div:nth-child(4) {
-      flex: 4;
-    }
-
-    & > div:last-child {
-      flex: 1;
-      align-items: flex-end;
-      padding-right: 0;
-    }
-
-    .hash {
-      max-width: 20vw;
-      min-width: 20vw;
-      display: inline-block;
-      overflow: hidden;
-    }
-  }
-}
-
-.elipsis {
-  max-width: 13vw;
-  display: inline-block;
-  overflow: hidden;
-}
-
-.tableRow {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: $spacing-base;
-  align-items: center;
-
-  & > div {
-    width: 100%;
-    display: flex;
-    justify-content: flex-start;
-    margin: 0 $spacing-base 0 $spacing-base;
-  }
-  & > div:nth-child(1) {
-    flex: 7;
-  }
-
-  & > div:nth-child(2) {
-    flex: 7;
-    overflow-wrap: anywhere;
-  }
-
-  & > div:nth-child(3) {
-    overflow: hidden;
-    flex: 10;
-  }
-  & > div:nth-child(4) {
-    flex: 7;
-  }
-  & > div:nth-child(5),
-  & > div:nth-child(6) {
-    flex: 3;
-  }
-}
-
-.hashEvents {
-  max-width: 13vw;
-  display: inline-block;
-  overflow: hidden;
-}
-
-.subTableValue {
-  max-width: 13vw;
-  display: inline-block;
-  overflow: hidden;
-  margin-right: $spacing-base;
-}
-
-.caretIcon {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
